feat(#1057): refactor species composition imports and add service for species composition management
